Revert "Feat 앱바 서브 메뉴 추가"
