--- apps/user/src/constants/nav-menu.ts
@@ -1,22 +1,0 @@
-export const NAV_MENUS = [
-  {
-    title: 'FORIF',
-    href: '/club/about',
-    submenu: [
-      { title: '포리프', href: '/club/about' },
-      { title: '포리프 팀', href: '/club/team' },
-    ],
-  },
-  {
-    title: '스터디',
-    href: '/studies',
-    submenu: [
-      { title: '스터디 목록', href: '/studies' },
-      { title: '스터디 지원', href: '/apply/member' },
-    ],
-  },
-  {
-    title: '자주 묻는 질문',
-    href: '/faq',
-  },
-];
