--- apps/user/src/routes/club/about.tsx
@@ -1,5 +1,0 @@
-import { createFileRoute } from '@tanstack/react-router'
-
-export const Route = createFileRoute('/club/about')({
-  component: () => <div>Hello /club/about!</div>
-})
