+++ packages/components/theme.ts
@@ -262,13 +262,6 @@ const getDesignTokens = (mode: 'light' | 'dark'): ThemeOptions => ({
         },
       },
     },
-    MuiAccordion: {
-      styleOverrides: {
-        root: {
-          padding: 0,
-        },
-      },
-    },
   },
   status: {
     danger: orange[500],
